# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: SHIELD API
  version: 1.0.0
  description: Smart Housing Infrastructure and Entry Log Digitalization API
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /staff/{id}:
    get:
      tags: [Staff]
      summary: Get staff
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Staff fetched}
    put:
      tags: [Staff]
      summary: Update staff
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Staff updated}
    delete:
      tags: [Staff]
      summary: Delete staff
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Staff deleted}
...
  /staff-attendance/staff/{staffId}:
    get:
      tags: [Staff]
      summary: List staff attendance by staff id
      parameters:
        - in: path
          name: staffId
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Staff attendance fetched}
...
  /staff-leaves:
    get:
      tags: [Staff]
      summary: List staff leaves
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Staff leaves fetched}
    post:
      tags: [Staff]
      summary: Create staff leave
      responses:
        '200': {description: Staff leave created}
  /staff-leaves/{id}:
    get:
      tags: [Staff]
      summary: Get staff leave
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Staff leave fetched}
    put:
      tags: [Staff]
      summary: Update staff leave
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Staff leave updated}
    delete:
      tags: [Staff]
      summary: Delete staff leave
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Staff leave deleted}
  /staff-leaves/{id}/approve:
    post:
      tags: [Staff]
      summary: Approve staff leave
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Staff leave approved}
  /staff-leaves/{id}/reject:
    post:
      tags: [Staff]
      summary: Reject staff leave
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Staff leave rejected}
  /staff-leaves/staff/{staffId}:
    get:
      tags: [Staff]
      summary: List staff leaves by staff
      parameters:
        - in: path
          name: staffId
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Staff leaves by staff fetched}
  /staff-leaves/pending-approval:
    get:
      tags: [Staff]
      summary: List pending staff leaves
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Pending staff leaves fetched}
  /staff-leaves/balance/{staffId}:
    get:
      tags: [Staff]
      summary: Get staff leave balance
      parameters:
        - in: path
          name: staffId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200': {description: Staff leave balance fetched}
  /payroll:
    get:
      tags: [Payroll]
      summary: List payroll
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Payroll fetched}
  /payroll/{id}:
    get:
      tags: [Payroll]
      summary: Get payroll
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Payroll fetched}
...
  /payroll/{id}/approve:
    post:
      tags: [Payroll]
      summary: Approve payroll
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Payroll approved}
  /payroll/{id}/payslip:
    get:
      tags: [Payroll]
      summary: Get payroll payslip
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Payroll payslip fetched}
  /payroll/month/{month}/year/{year}:
    get:
      tags: [Payroll]
      summary: List payroll by period
      parameters:
        - in: path
          name: month
          required: true
          schema:
            type: integer
        - in: path
          name: year
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Payroll by period fetched}
  /payroll/staff/{staffId}:
    get:
      tags: [Payroll]
      summary: List payroll by staff
      parameters:
        - in: path
          name: staffId
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200': {description: Payroll by staff fetched}
  /payroll/summary:
    get:
      tags: [Payroll]
      summary: Payroll summary
      parameters:
        - in: query
          name: month
          required: false
          schema:
            type: integer
        - in: query
          name: year
          required: false
          schema:
            type: integer
      responses:
        '200': {description: Payroll summary fetched}
  /payroll-components:
    get:
      tags: [Payroll]
      summary: List payroll components
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200': {description: Payroll components fetched}
    post:
      tags: [Payroll]
      summary: Create payroll component
      responses:
        '200': {description: Payroll component created}
  /payroll-components/{id}:
    get:
      tags: [Payroll]
      summary: Get payroll component
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Payroll component fetched}
    put:
      tags: [Payroll]
      summary: Update payroll component
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Payroll component updated}
    delete:
      tags: [Payroll]
      summary: Delete payroll component
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': {description: Payroll component deleted}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Sort:
      in: query
      name: sort
      schema:
        type: string
        example: createdAt,desc
...